feature: refactor signin sequence with code scan option
