+++ src/api/mock/index.ts
@@ -4,9 +4,9 @@ export const apiBaseUrl = import.meta.env.VITE_BOOK_API_BASE_URL;
 
 export const getAllBooks = () => {
     return fetch(apiBaseUrl)
-        .then(resp => resp.json())
-        .then(jsonResp => (jsonResp as unknown) as Book[]);
-}
+        .then((resp) => resp.json())
+        .then((jsonResp) => jsonResp as unknown as Book[]);
+};
 
 export const deleteBookById = (id: string): Promise<Response> =>
     fetch(`${apiBaseUrl}/${id}`, {
@@ -32,17 +32,26 @@ export const updateBook = (book: Book): Promise<Book> =>
         }),
     }).then((resp) => resp.json() as unknown as Book);
 
-export const login = (name:string, password:string) => new Promise( (resolve, reject) => {
-    setTimeout(() => {
-        resolve('dummy_key');
-    }, 1000);
-});
+export const login = (name: string, password: string) =>
+    new Promise((resolve, reject) => {
+        setTimeout(() => {
+            resolve("dummy_key");
+        }, 1000);
+    });
 
-export const logout = () => new Promise( (resolve, reject) => {
-    setTimeout(() => {
-        resolve(true);
-    }, 1000);
-});
+export const logout = () =>
+    new Promise((resolve, reject) => {
+        setTimeout(() => {
+            resolve(true);
+        }, 1000);
+    });
+
+export const checkApiKey = () =>
+    new Promise<boolean>((resolve, reject) => {
+        setTimeout(() => {
+            resolve(true);
+        }, 1000);
+    });
 
 export default {
     getAllBooks,
@@ -50,5 +59,6 @@ export default {
     addBook,
     updateBook,
     login,
-    logout
+    logout,
+    checkApiKey
 };
