+++ src/api/private/index.ts
@@ -1,4 +1,4 @@
-import { Book, LoginSuccessResponse } from "../../types";
+import { Book, LoginSuccessResponse, ApiKeyCheckResponse } from "../../types";
 import { nanoid } from "nanoid";
 import Storage from "@/utils/storage";
 
@@ -10,7 +10,7 @@ const getApiKey = () => Storage.getApiKey() || "";
 const handleErrorJson = (response: Response) => {
     if (!response.ok) {
         return response.json().then((data) => {
-            throw data
+            throw data;
         });
     } else {
         return response.json();
@@ -19,7 +19,7 @@ const handleErrorJson = (response: Response) => {
 
 const handleErrorNoResponse = (response: Response) => {
     if (!response.ok) {
-        throw response.statusText;        
+        throw response.statusText;
     } else {
         return true;
     }
@@ -37,7 +37,7 @@ export const getAllBooks = () =>
         .then(handleErrorJson)
         .then((jsonResp) => jsonResp as unknown as Book[]);
 
-export const addBook = (book: Omit<Book, 'id'>): Promise<Book> =>
+export const addBook = (book: Omit<Book, "id">): Promise<Book> =>
     fetch(
         `${apiBaseUrl}?${new URLSearchParams({
             r: "api/user-book/create",
@@ -123,6 +123,20 @@ export const logout = () =>
         }, 100);
     });
 
+export const checkApiKey = (apiKey: string) =>
+    fetch(
+        `${apiBaseUrl}?${new URLSearchParams({
+            r: "api/auth/check-api-key",
+            token: apiKey,
+        })}`
+    )
+        .then(handleErrorJson)
+        .then((successResponse) => {
+            const apiKeyCheck =
+                successResponse as unknown as ApiKeyCheckResponse;
+            return apiKeyCheck.isValid;
+        });
+
 export default {
     getAllBooks,
     deleteBookById,
@@ -130,4 +144,5 @@ export default {
     updateBook,
     login,
     logout,
+    checkApiKey
 };
