+++ src/component/WithNavigation.tsx
@@ -10,7 +10,6 @@ import { SettingsPage } from "../page/SettingsPage";
 
 import Storage from "../utils/storage";
 import { SignInPage } from "@/page/SignInPage";
-import { ApiKeyScannerPage } from "@/page/ApiKeyScannerPage";
 
 export const WithNavigation: React.FC<{}> = (): JSX.Element => {
     return (
@@ -19,7 +18,6 @@ export const WithNavigation: React.FC<{}> = (): JSX.Element => {
             <Route path="/about" component={AboutPage} />
             <Route path="/add" component={AddBookPage} />
             <Route path="/settings" component={SettingsPage} />
-            <Route path="/scan-signin" component={ApiKeyScannerPage} />
             <Route path="/update/:id">
                 {(params) => <UpdateBookPage id={params.id} />}
             </Route>
