+++ src/types.ts
@@ -29,7 +29,11 @@ export type LoginSuccessResponse = {
      * api key revoked)
      */
     api_key?: string;
-}
+};
+
+export type ApiKeyCheckResponse = {
+    isValid: boolean;
+};
 
 export type ErrorResponse = {
     name: string;
